Restructure heap
